+++ libs/spark/src/useDropdown_unstable/useDropdown_unstable.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import Unstable_DropdownContext, {
+  Unstable_DropdownContextValue,
+} from '../Unstable_DropdownContext';
+
+const useDropdown_unstable = (): Unstable_DropdownContextValue => {
+  return useContext(Unstable_DropdownContext);
+};
+
+export default useDropdown_unstable;